rgw/lua: replace int64_t with "long long int" for the background table
lua integer is "long long int" which works with GCC 11.2.1 (fedora 36)
but fails to compile on centos8 and ubuntu focal

+++ src/test/rgw/test_rgw_lua.cc
@@ -877,7 +877,7 @@ TEST(TestRGWLuaBackground, TableValues)
   const auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<double>(lua_background, "key3"), 42.2);
   EXPECT_TRUE(get_table_value<bool>(lua_background, "key4"));
 }
@@ -897,7 +897,7 @@ TEST(TestRGWLuaBackground, TablePersist)
   auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   
   request_script = R"(
     RGW["key3"] = RGW["key1"]
@@ -907,9 +907,9 @@ TEST(TestRGWLuaBackground, TablePersist)
   rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key3"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key4"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key4"), 42);
 }
 
 TEST(TestRGWLuaBackground, TableValuesFromRequest)
@@ -934,7 +934,7 @@ TEST(TestRGWLuaBackground, TableValuesFromRequest)
 
   const auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key1"), -99);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key1"), -99);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key2"), "hi");
   EXPECT_EQ(get_table_value<double>(lua_background, "key3"), -9.9);
   EXPECT_EQ(get_table_value<bool>(lua_background, "key4"), true);
@@ -961,7 +961,7 @@ TEST(TestRGWLuaBackground, TableInvalidValue)
   const auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_NE(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "val1");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<double>(lua_background, "key3"), 42.2);
   EXPECT_EQ(get_table_value<bool>(lua_background, "key4"), true);
 }
@@ -984,9 +984,9 @@ TEST(TestRGWLuaBackground, TableErase)
   auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key3"), "another string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "size"), 4);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "size"), 4);
   
   request_script = R"(
     -- erase key1
@@ -999,9 +999,9 @@ TEST(TestRGWLuaBackground, TableErase)
   rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key3"), "another string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "size"), 3);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "size"), 3);
 }
 
 TEST(TestRGWLuaBackground, TableIterate)
@@ -1025,10 +1025,10 @@ TEST(TestRGWLuaBackground, TableIterate)
   const auto rc = lua::request::execute(nullptr, nullptr, nullptr, &s, "", request_script);
   ASSERT_EQ(rc, 0);
   EXPECT_EQ(get_table_value<std::string>(lua_background, "key1"), "string value");
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "key2"), 42);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "key2"), 42);
   EXPECT_EQ(get_table_value<double>(lua_background, "key3"), 42.2);
   EXPECT_TRUE(get_table_value<bool>(lua_background, "key4"));
-  EXPECT_EQ(get_table_value<int64_t>(lua_background, "size"), 5);
+  EXPECT_EQ(get_table_value<long long int>(lua_background, "size"), 5);
 }
 
 TEST(TestRGWLuaBackground, TableIncrement)
